requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
boto3==1.17.76
botocore==1.20.76
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
multidict==5.1.0
pyee==7.0.4
pyslack==0.5.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
requests==2.25.1
s3transfer==0.4.2
schedule==1.1.0
six==1.16.0
slack==0.0.2
slack-bolt==1.6.0
slack-sdk==3.5.1
slackclient==1.3.2
slackeventsapi==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.4
websocket-client==0.54.0
Werkzeug==1.0.1
yarl==1.6.3